Sokol buffers added
Added the description buffer, vertex buffer and the pipeline buffer to the graphics manager. Placed within the private implementation class so that they can remain hidden from the user. May have to add setters to allow the user to create their own structures and pass them to the graphics manager.